"Moving around the sidebar file back to docs/ but with ./sidebar.js relative import still..."
