+++ package.json
@@ -8,5 +8,8 @@
     "vitepress": "^1.0.0-beta.1",
     "vitepress-plugin-tabs": "^0.3.0"
   },
-  "type": "module"
+  "type": "module",
+  "dependencies": {
+    "sidebar": "^1.0.2"
+  }
 }
